requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
gunicorn==20.1.0
Werkzeug==2.2.3
